apps/briefs/requirements.txt
flask
pydantic
numpy
pandas
scikit-learn
umap-learn
hdbscan
transformers>=4.48.0
torch
tqdm
python-dotenv
requests
retry
openai
json-repair
google-genai
gliclass
rapidfuzz
ipywidgets
widgetsnbextension
pandas-profiling
readtime
optuna
bertopic
supabase